+++ pyabc/epsilon/epsilon.py
@@ -138,6 +138,7 @@ class QuantileEpsilon(Epsilon):
     def initialize(self,
                    t: int,
                    get_weighted_distances: Callable[[], pd.DataFrame],
+                   get_all_records: Callable[[], List[dict]],
                    max_nr_populations: int,
                    acceptor_config: dict):
         if self._initial_epsilon != 'from_sample':
@@ -181,14 +182,18 @@ class QuantileEpsilon(Epsilon):
 
     def update(self,
                t: int,
-               weighted_distances: pd.DataFrame,
+               get_weighted_distances: Callable[[], pd.DataFrame],
+               get_all_records: Callable[[], List[dict]],
                acceptance_rate: float,
                acceptor_config: dict):
         """
         Compute quantile of the (weighted) distances given in population,
         and use this to update epsilon.
         """
+        # execute function
+        weighted_distances = get_weighted_distances()
 
+        # update epsilon
         self._update(t, weighted_distances)
 
         # logger